Update ldap.php
Installed the plugin so that non-windows, email-only users could update their passwords but wanted the windows users to be able to too. The 'Update Samba' stuff causes roundcube to claim mistakenly that the password write was unsuccessful if it's a posix-only user. my patch fixes that by making both smbpwattr and smblchattr to be false, which then bypasses the 'Update Samba' calls.